Add cast to float

+++ flamedisx/core.py
@@ -380,7 +380,8 @@ class ERSource:
         q_produced = tf.floor(es / self.gimme('work')[:, o])
 
         # (n_events, |nq|, |ne|) tensor giving p(nq | e)
-        p_nq_e = tf.equal(nq_1d[:, :, o], q_produced[:, o, :])
+        p_nq_e = tf.cast(tf.equal(nq_1d[:, :, o], q_produced[:, o, :]),
+                         dtype=tf.float64)
 
         return tf.reduce_sum(p_nq_e * rate_e[:, o, :], axis=2)
